Cleaned up testcase, merged with other SpanByte test.
